+++ offlineimap/utils/distro_utils.py
